Added meta templates for pyrus

+++ build/scripts/apidoc-release.php
@@ -13,6 +13,11 @@ if (!$release) {
 $apidoc_package     = ROOT . '/packages/apidoc/ZendFramework-' . $release . '-apidoc.zip';
 $apidoc_release_tmp = ROOT . '/public/docs/ZendFramework-' . $release . '/tmp';
 $apidoc_release_dir = ROOT . '/public/docs/ZendFramework-' . $release . '/apidoc';
+
+if (!file_exists($apidoc_package)) {
+    exit(0);
+}
+
 mkdir($apidoc_release_tmp, 0777, true);
 script_run_command('unzip ' . $apidoc_package . ' -d ' . $apidoc_release_tmp);
 script_run_command('mv ' . $apidoc_release_tmp . '/ZendFramework-' . $release . ' ' . $apidoc_release_dir);
